+++ test/db.js
@@ -40,7 +40,7 @@ buster.testCase('db - createDatabases', {
       nano: this._mockNano(function (dbName, cb) { cb(null, { ok: 'true' }); })
     });
     db.createDatabases(['db1', 'db2'], function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 2);
       assert.equals(result[0].id, 'db1');
       assert.equals(result[0].message, 'created');
@@ -54,7 +54,7 @@ buster.testCase('db - createDatabases', {
       nano: this._mockNano(function (dbName, cb) { cb({ status_code: 412, error: 'file_exists' }); })
     });
     db.createDatabases(['db1', 'db2'], function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 2);
       assert.equals(result[0].id, 'db1');
       assert.equals(result[0].message, 'file_exists (412)');
@@ -78,7 +78,7 @@ buster.testCase('db - removeDatabases', {
       nano: this._mockNano(function (dbName, cb) { cb(null, { ok: 'true' }); })
     });
     db.removeDatabases(['db1', 'db2'], function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 2);
       assert.equals(result[0].id, 'db1');
       assert.equals(result[0].message, 'deleted');
@@ -92,7 +92,7 @@ buster.testCase('db - removeDatabases', {
       nano: this._mockNano(function (dbName, cb) { cb({ status_code: 404, error: 'missing' }); })
     });
     db.removeDatabases(['db1', 'db2'], function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 2);
       assert.equals(result[0].id, 'db1');
       assert.equals(result[0].message, 'missing (404)');
@@ -163,7 +163,7 @@ buster.testCase('db - createDocuments', {
       nano: this._mockNano(function (doc, cb) { cb(null, { ok: 'true' }); })
     });
     db.createDocuments({ db1: [{ _id: 'docA' }], db2: [{ _id: 'docB'}, { _id: 'docC'}] }, function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 3);
       assert.equals(result[0].id, 'db1/docA');
       assert.equals(result[0].message, 'created');
@@ -179,7 +179,7 @@ buster.testCase('db - createDocuments', {
       nano: this._mockNano(function (doc, cb) { cb({ status_code: 409, error: 'conflict' }); })
     });
     db.createDocuments({ db1: [{ _id: 'docA' }], db2: [{ _id: 'docB'}, { _id: 'docC'}] }, function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 3);
       assert.equals(result[0].id, 'db1/docA');
       assert.equals(result[0].message, 'conflict (409)');
@@ -205,7 +205,7 @@ buster.testCase('db - saveDocuments', {
       nano: this._mockNano(function (doc, cb) { cb(null, { ok: 'true' }); })
     });
     db.saveDocuments({ db1: [{ _id: 'docA' }], db2: [{ _id: 'docB'}, { _id: 'docC'}] }, function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 3);
       assert.equals(result[0].id, 'db1/docA');
       assert.equals(result[0].message, 'created');
@@ -232,7 +232,7 @@ buster.testCase('db - saveDocuments', {
       )
     });
     db.saveDocuments({ db1: [{ _id: 'docA' }], db2: [{ _id: 'docB'}, { _id: 'docC'}] }, function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 3);
       assert.equals(result[0].id, 'db1/docA');
       assert.equals(result[0].message, 'missing (404)');
@@ -259,7 +259,7 @@ buster.testCase('db - saveDocuments', {
       )
     });
     db.saveDocuments({ db1: [{ _id: 'docA' }], db2: [{ _id: 'docB'}, { _id: 'docC'}] }, function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 3);
       assert.equals(result[0].id, 'db1/docA');
       assert.equals(result[0].message, 'updated');
@@ -292,7 +292,7 @@ buster.testCase('db - removeDocuments', {
       )
     });
     db.removeDocuments({ db1: [{ _id: 'docA' }], db2: [{ _id: 'docB'}, { _id: 'docC'}] }, function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 3);
       assert.equals(result[0].id, 'db1/docA');
       assert.equals(result[0].message, 'deleted');
@@ -312,7 +312,7 @@ buster.testCase('db - removeDocuments', {
       )
     });
     db.removeDocuments({ db1: [{ _id: 'docA' }], db2: [{ _id: 'docB'}, { _id: 'docC'}] }, function (err, result) {
-      assert.isNull(err);
+      assert.equals(err, undefined);
       assert.equals(result.length, 3);
       assert.equals(result[0].id, 'db1/docA');
       assert.equals(result[0].message, 'missing (404)');
@@ -358,7 +358,7 @@ buster.testCase('db - warmViews', {
       });
     db.warmViews({ db1: [{ _id: 'docA' }, { _id: '_design/somedesigndoc', views: { view1: {}, view2: {} } } ] }, 
      function (err, result) {
-       assert.isNull(err);
+       assert.equals(err, undefined);
        assert.equals(result.length, 1);
        assert.equals(result[0].id, 'db1/_design/somedesigndoc');
        assert.equals(result[0].message, 'warmed view1, view2');
@@ -390,10 +390,11 @@ buster.testCase('db - liveDeployView', {
   'should ignore non design docs': function (done) {
     var db = new Db('http://localhost:5984', {
       nano: this._mockNano(),
-      interval: 0
+      interval: 0,
+      parallelLimit: 100
     });
     db.liveDeployView({ db1: [{ _id: 'docA' }]}, function (err, result) {
-      assert.equals(err, null);
+      assert.equals(err, undefined);
       assert.isTrue(_.isEmpty(result.length));
       done();
     });
@@ -465,7 +466,7 @@ buster.testCase('db - liveDeployView', {
         ] 
       }, 
      function (err, result) {
-       assert.isNull(err);
+       assert.equals(err, undefined);
        assert.isFalse(_.isEmpty(result));
        assert.equals('view index has been kicked off', result[0][1][0].message);
        assert.equals('view index has been kicked off', result[0][1][1].message);
@@ -539,7 +540,7 @@ buster.testCase('db - liveDeployView', {
         ] 
       }, 
      function (err, result) {
-       assert.isNull(err);
+       assert.equals(err, undefined);
        assert.isTrue(!_.isEmpty(result));
        assert.equals("success", result[0][3][0].message);
        assert.equals("success", result[0][3][1].message);
@@ -613,7 +614,7 @@ buster.testCase('db - liveDeployView', {
         ] 
       }, 
      function (err, result) {
-       assert.isNull(err);
+       assert.equals(err, undefined);
        assert.isTrue(!_.isEmpty(result));
        assert.isTrue(result[0][4].ok);
        done();
@@ -668,7 +669,7 @@ buster.testCase('db - liveDeployView', {
         ] 
       }, 
      function (err, result) {
-       assert.isNull(err);
+       assert.equals(err, undefined);
        assert.isTrue(!_.isEmpty(result));
        assert.isTrue(result[0][4].ok);
        done();